feat(Menu): Add customisable menu component
