+++ lib/components/OverflowMenu/OverflowMenu.docs.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ComponentDocs } from '../../../site/src/types';
-import { Box, OverflowMenu, OverflowMenuItem } from '../';
+import { Box, OverflowMenu, MenuItem } from '../';
 
 const docs: ComponentDocs = {
   category: 'Interaction',
@@ -11,17 +11,8 @@ const docs: ComponentDocs = {
       Example: ({ handler }) => (
         <Box style={{ paddingLeft: '100px', maxWidth: '200px' }}>
           <OverflowMenu label="Options">
-            <OverflowMenuItem onClick={handler}>First</OverflowMenuItem>
-            <OverflowMenuItem onClick={handler}>Second</OverflowMenuItem>
-            <OverflowMenuItem type="link">
-              <Box
-                component="a"
-                href="https://seek-oss.github.io/braid-design-system/components/OverflowMenu"
-                target="_blank"
-              >
-                Link to page
-              </Box>
-            </OverflowMenuItem>
+            <MenuItem onClick={handler}>First</MenuItem>
+            <MenuItem onClick={handler}>Second</MenuItem>
           </OverflowMenu>
         </Box>
       ),
@@ -32,8 +23,8 @@ const docs: ComponentDocs = {
       name: 'Standard',
       code: (
         <OverflowMenu label="Options">
-          <OverflowMenuItem>Option</OverflowMenuItem>
-          <OverflowMenuItem>Option</OverflowMenuItem>
+          <MenuItem>Option</MenuItem>
+          <MenuItem>Option</MenuItem>
         </OverflowMenu>
       ),
     },
