+++ lib/components/OverflowMenu/OverflowMenu.treat.ts
@@ -1,5 +1,4 @@
 import { style } from 'sku/treat';
-import * as zIndex from '../private/zIndex';
 
 const OVERFLOW_ICON_WIDTH = 4;
 const OVERFLOW_ICON_HEIGHT = 14;
@@ -11,31 +10,5 @@ export const triggerOffset = style(theme => {
 
   return {
     margin: `-${offsetY}px -${offsetX}px`,
-    // Minusing one here to avoid buttons overriding open menus below
-    zIndex: zIndex.pageOverlay - 1,
   };
 });
-
-export const root = style({
-  padding: '0.05px',
-  lineHeight: 0,
-});
-
-export const backdrop = style({
-  width: '100vw',
-  height: '100vh',
-  zIndex: zIndex.backdrop,
-});
-
-export const menuIsClosed = style(theme => ({
-  transform: `translateY(-${theme.grid * 2}px)`,
-  opacity: 0,
-  visibility: 'hidden',
-}));
-
-export const menu = style({
-  zIndex: zIndex.pageOverlay,
-  right: 0,
-});
-
-export const showOverlay = style({ opacity: 1 });
