+++ lib/components/index.ts
@@ -28,8 +28,9 @@ export { Heading } from './Heading/Heading';
 export { Hidden } from './Hidden/Hidden';
 export { Inline } from './Inline/Inline';
 export { Loader } from './Loader/Loader';
+export { Menu } from './Menu/Menu';
+export { MenuItem } from './MenuItem/MenuItem';
 export { OverflowMenu } from './OverflowMenu/OverflowMenu';
-export { OverflowMenuItem } from './OverflowMenuItem/OverflowMenuItem';
 export { MonthPicker } from './MonthPicker/MonthPicker';
 export { Notice } from './Notice/Notice';
 export { Radio } from './Radio/Radio';
