+++ site/src/undocumentedExports.json
@@ -1,1 +1,1 @@
-["BoxRenderer", "ToastProvider"]
+["BoxRenderer", "ToastProvider", "MenuItem"]
